Rename files so they follow PSR-4
