+++ example.php
@@ -1,9 +1,9 @@
 <?php
 
   // Require source files (not needed if you use composer)
-  require 'src/crunchdb.class.php';
-  require 'src/crunchtable.class.php';
-  require 'src/crunchresource.class.php';
+  require 'src/crunchDB.php';
+  require 'src/crunchTable.php';
+  require 'src/crunchResource.php';
 
   // Alias class accessor to omit namespace (optional)
   use cybrox\crunchdb\crunchDB as crunchDB;
